update: 体温をビーコンに乗せて送信完了
